# api.yaml
swagger: '2.0'
info:
  description: cruise_management system restful api
  version: '1.0'
  title: cruise_management
host: 'localhost:8083'
basePath: /
...
paths:
...
  /records/create:
    post:
      tags:
        - record-controller
      summary: list
      operationId: listUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: bigShipTraffic
          in: query
          description: maximum traffic of big ship
          required: false
          type: integer
          format: int32
          x-example: 200
        - name: date
          in: query
          description: date of the day
          required: false
          type: string
          x-example: '2019-02-18'
        - name: endTime
          in: query
          description: end time
          required: false
          type: string
          x-example: '07:15:00'
        - name: isGoldenWeek
          in: query
          description: is golden week or not
          required: false
          type: boolean
          x-example: 'false'
        - name: smallShipTraffic
          in: query
          description: maximum traffic of small ship
          required: false
          type: integer
          format: int32
          x-example: 100
        - name: startTime
          in: query
          description: start time
          required: false
          type: string
          x-example: '07:00:00'
        - name: timeForSailing
          in: query
          description: flight time
          required: false
          type: integer
          format: int32
          x-example: '15'
        - name: traffic
          in: query
          description: forecasting traffic
          required: false
          type: integer
          format: int32
          x-example: 10040
        - name: weather
          in: query
          description: weather
          required: false
          type: string
          x-example: good weather
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResultVO«RecordsVO»'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /records/delete:
    get:
      tags:
        - record-controller
      summary: list records by id
      description: list records by id
      operationId: deleteUsingGET
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: id
          in: query
          description: id
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResultVO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /records/list:
    get:
      tags:
        - record-controller
      summary: list records by date
      description: list records by date
      operationId: listUsingGET_1
      produces:
        - '*/*'
        - application/json;charset=UTF-8
      parameters:
        - name: date
          in: query
          description: date
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ResultVO«RecordDto»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
...
definitions:
...
  RecordDto:
    type: object
    properties:
      arriveBerth:
        type: string
      arriveBerthTime:
        type: string
      arriveTime:
        type: string
      id:
        type: string
      shipName:
        type: string
      startBerth:
        type: string
      startTime:
        type: string
      timeForSailing:
        type: integer
        format: int32
      waitTime:
        type: integer
        format: int32
    title: RecordDto
  RecordsVO:
    type: object
    properties:
      bigShipNumber:
        type: integer
        format: int32
      bigShipTraffic:
        type: integer
        format: int32
      bigShipWaitTime:
        type: integer
        format: int32
      date:
        type: string
      hours:
        type: string
      isGoldenWeek:
        type: boolean
      records:
        type: array
        items:
          $ref: '#/definitions/RecordDto'
      smallShipNumber:
        type: integer
        format: int32
      smallShipTraffic:
        type: integer
        format: int32
      smallShipWaitTime:
        type: integer
        format: int32
      timeForSailing:
        type: integer
        format: int32
      traffic:
        type: integer
        format: int32
      weather:
        type: string
    title: RecordsVO
  ResultVO:
    type: object
    properties:
      code:
        type: integer
        format: int32
      data:
        type: object
      msg:
        type: string
    title: ResultVO
...
  ResultVO«RecordsVO»:
    type: object
    properties:
      code:
        type: integer
        format: int32
      data:
        $ref: '#/definitions/RecordsVO'
      msg:
        type: string
    title: ResultVO«RecordsVO»
...